requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
Pillow==11.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.4

# Google OAuth dependencies
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
requests-oauthlib==2.0.0
